Convert student profile to dedicated screen
