+++ app/src/main/java/com/tutorly/ui/screens/StudentProfileUiState.kt
@@ -1,0 +1,10 @@
+package com.tutorly.ui.screens
+
+import com.tutorly.domain.model.StudentProfile
+
+sealed interface StudentProfileUiState {
+    data object Hidden : StudentProfileUiState
+    data object Loading : StudentProfileUiState
+    data object Error : StudentProfileUiState
+    data class Content(val profile: StudentProfile) : StudentProfileUiState
+}
